Agentic-Multidomain-Reasoner/src/tools/z3_tool.py
# src/tools/z3_tool.py
"""Z3 SMT solver tool for constraint satisfaction with optional dependency."""
from typing import Dict, Any

class Z3Tool:
    def execute(self, operation: str = "check_sat", **kwargs) -> Dict[str, Any]:
        try:
            from z3 import Solver, Int, Real, Bool, sat  # lazy import
        except Exception as e:
            return {"success": False, "data": {}, "error": f"Z3Unavailable: {e}"}
        try:
            if operation == "check_sat":
                solver = Solver()
                # Minimal placeholder: accept pre-encoded constraints if provided
                constraints = kwargs.get("constraints", [])
                for c in constraints:
                    solver.add(c)
                res = solver.check()
                return {"success": True, "data": {"result": str(res)}, "error": ""}
            else:
                return {"success": False, "data": {}, "error": f"Unknown operation: {operation}"}
        except Exception as e:
            return {"success": False, "data": {}, "error": f"{type(e).__name__}: {str(e)}"}
