Agentic-Multidomain-Reasoner/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-reasoning-system"
version = "0.1.0"
description = "Agentic AI reasoning system for logic puzzles and MCQs"
authors = [{name = "Saptang Labs", email = "team@saptanglabs.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    "sympy>=1.12",
    "z3-solver>=4.12.2",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "simhash>=2.1.2",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "jsonschema>=4.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
